production database, changed table columns

+++ db/migrate/20191019144438_create_services.rb
@@ -2,7 +2,7 @@ class CreateServices < ActiveRecord::Migration[6.0]
   def change
     create_table :services do |t|
       t.string :serviceName
-      t.decimal :servicePrice
+      t.integer :servicePrice
 
       t.timestamps
     end
