+++ db/migrate/20191019151217_create_app_durations.rb
@@ -2,7 +2,7 @@ class CreateAppDurations < ActiveRecord::Migration[6.0]
   def change
     create_table :app_durations do |t|
       t.integer :duration
-      t.decimal :durationPrice
+      t.integer :durationPrice
 
       t.timestamps
     end
